misc: 独立wbiSign代码，修改缓存判断
